+++ sdk/appconfiguration/Azure.Data.AppConfiguration/src/Models/SnapshotFields.cs
@@ -22,41 +22,38 @@ namespace Azure.Data.AppConfiguration
 
         private const string NameValue = "name";
         private const string StatusValue = "status";
-        private const string StatusCodeValue = "status_code";
         private const string FiltersValue = "filters";
-        private const string CompositionTypeValue = "composition_type";
-        private const string CreatedValue = "created";
-        private const string ExpiresValue = "expires";
+        private const string SnapshotCompositionValue = "composition_type";
+        private const string CreatedOnValue = "created";
+        private const string ExpiresOnValue = "expires";
         private const string RetentionPeriodValue = "retention_period";
-        private const string SizeValue = "size";
-        private const string ItemsCountValue = "items_count";
+        private const string SizeInBytesValue = "size";
+        private const string ItemCountValue = "items_count";
         private const string TagsValue = "tags";
-        private const string EtagValue = "etag";
+        private const string ETagValue = "etag";
 
         /// <summary> name. </summary>
         public static SnapshotFields Name { get; } = new SnapshotFields(NameValue);
         /// <summary> status. </summary>
         public static SnapshotFields Status { get; } = new SnapshotFields(StatusValue);
-        /// <summary> status_code. </summary>
-        public static SnapshotFields StatusCode { get; } = new SnapshotFields(StatusCodeValue);
         /// <summary> filters. </summary>
         public static SnapshotFields Filters { get; } = new SnapshotFields(FiltersValue);
         /// <summary> composition_type. </summary>
-        public static SnapshotFields CompositionType { get; } = new SnapshotFields(CompositionTypeValue);
+        public static SnapshotFields SnapshotComposition { get; } = new SnapshotFields(SnapshotCompositionValue);
         /// <summary> created. </summary>
-        public static SnapshotFields Created { get; } = new SnapshotFields(CreatedValue);
+        public static SnapshotFields CreatedOn { get; } = new SnapshotFields(CreatedOnValue);
         /// <summary> expires. </summary>
-        public static SnapshotFields Expires { get; } = new SnapshotFields(ExpiresValue);
+        public static SnapshotFields ExpiresOn { get; } = new SnapshotFields(ExpiresOnValue);
         /// <summary> retention_period. </summary>
         public static SnapshotFields RetentionPeriod { get; } = new SnapshotFields(RetentionPeriodValue);
         /// <summary> size. </summary>
-        public static SnapshotFields Size { get; } = new SnapshotFields(SizeValue);
+        public static SnapshotFields SizeInBytes { get; } = new SnapshotFields(SizeInBytesValue);
         /// <summary> items_count. </summary>
-        public static SnapshotFields ItemsCount { get; } = new SnapshotFields(ItemsCountValue);
+        public static SnapshotFields ItemCount { get; } = new SnapshotFields(ItemCountValue);
         /// <summary> tags. </summary>
         public static SnapshotFields Tags { get; } = new SnapshotFields(TagsValue);
         /// <summary> etag. </summary>
-        public static SnapshotFields Etag { get; } = new SnapshotFields(EtagValue);
+        public static SnapshotFields ETag { get; } = new SnapshotFields(ETagValue);
         /// <summary> Determines if two <see cref="SnapshotFields"/> values are the same. </summary>
         public static bool operator ==(SnapshotFields left, SnapshotFields right) => left.Equals(right);
         /// <summary> Determines if two <see cref="SnapshotFields"/> values are not the same. </summary>