+++ sdk/digitaltwins/Azure.DigitalTwins.Core/src/Customized/Models/ImportJobStatus.cs
@@ -1,0 +1,17 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+using System;
+using Azure.Core;
+
+namespace Azure.DigitalTwins.Core
+{
+    /// <summary>
+    /// Status of Import Job
+    /// </summary>
+    [CodeGenModel("Status")]
+    public readonly partial struct ImportJobStatus
+    {
+        // This class delcaration sets the name space and sets the enum to be extensible; do not remove.
+    }
+}