+++ sdk/managednetworkfabric/Azure.ResourceManager.ManagedNetworkFabric/tests/Scenario/IpExtendedCommunityTests.cs
@@ -64,6 +64,20 @@ namespace Azure.ResourceManager.ManagedNetworkFabric.Tests.Scenario
             }
             Assert.IsNotEmpty(listByResourceGroup);
 
+            //List by subscription
+            ResourceIdentifier subscriptionResourceId = SubscriptionResource.CreateResourceIdentifier(TestEnvironment.SubscriptionId);
+            SubscriptionResource subscriptionResource = Client.GetSubscriptionResource(subscriptionResourceId);
+
+            TestContext.Out.WriteLine($"GET - List by Subscription started.....");
+
+            await foreach (NetworkFabricIPExtendedCommunityResource item in subscriptionResource.GetNetworkFabricIPExtendedCommunitiesAsync())
+            {
+                NetworkFabricIPExtendedCommunityData resourceData = item.Data;
+                TestContext.WriteLine($"Succeeded on id: {resourceData.Id}");
+            }
+
+            TestContext.Out.WriteLine($"List by Subscription operation succeeded.");
+
             // Delete
             TestContext.Out.WriteLine($"DELETE started.....");
             var deleteResponse = await ipExtendedCommunity.DeleteAsync(WaitUntil.Completed);