+++ sdk/managednetworkfabric/Azure.ResourceManager.ManagedNetworkFabric/tests/Scenario/NetworkDeviceTests.cs
@@ -48,6 +48,20 @@ namespace Azure.ResourceManager.ManagedNetworkFabric.Tests.Scenario
             }
             Assert.IsNotEmpty(listByResourceGroup);
 
+            //List by subscription
+            ResourceIdentifier subscriptionResourceId = SubscriptionResource.CreateResourceIdentifier(TestEnvironment.SubscriptionId);
+            SubscriptionResource subscriptionResource = Client.GetSubscriptionResource(subscriptionResourceId);
+
+            TestContext.Out.WriteLine($"GET - List by Subscription started.....");
+
+            await foreach (NetworkDeviceResource item in subscriptionResource.GetNetworkDevicesAsync())
+            {
+                NetworkDeviceData resourceData = item.Data;
+                TestContext.WriteLine($"Succeeded on id: {resourceData.Id}");
+            }
+
+            TestContext.Out.WriteLine($"List by Subscription operation succeeded.");
+
             // Update Serial Number
             NetworkDevicePatch patch = new NetworkDevicePatch()
             {