+++ sdk/storage/Azure.Storage.Files.DataLake/tests/DataLakeClientTestFixtureAttribute.cs
@@ -28,6 +28,7 @@ namespace Azure.Storage.Files.DataLake.Tests
                 DataLakeClientOptions.ServiceVersion.V2023_01_03,
                 DataLakeClientOptions.ServiceVersion.V2023_05_03,
                 DataLakeClientOptions.ServiceVersion.V2023_08_03,
+                DataLakeClientOptions.ServiceVersion.V2023_11_03,
                 StorageVersionExtensions.LatestVersion,
                 StorageVersionExtensions.MaxVersion)
         {