+++ sdk/storage/Microsoft.Azure.WebJobs.Extensions.Storage.Queues/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 5.2.0-beta.1 (Unreleased)
+## 5.3.0-beta.1 (Unreleased)
 
 ### Features Added
 
@@ -9,6 +9,10 @@
 ### Bugs Fixed
 
 ### Other Changes
+
+## 5.2.0 (2023-09-25)
+### Other Changes
+- This release contains bug fixes to improve quality.
 
 ## 5.1.3 (2023-06-26)
 - Trigger binding support for ParameterBindingData reference type